requirements.txt
azure-appconfiguration==1.4.0
azure-appconfiguration-provider==1.0.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
